Seattle AirBnb's dataset using Tableau
Using the information from a dataset of the Airbnb's in Seattle to create useful interpretations of the data for the prospective buyer in mind. In order to understand the average price within a zip code, the average price per bedrooms, the count of listings per number of bedrooms and the revenue at throughout the year of 2016, thus; understanding which time period is more profitable. The information is displayed as an interactive dashboard.
The data is cleaned whilst within an Excel spreadsheet then transferred to Tableau to transform the data in to useful charts and graphs for the dashboard.