Amazon Web Scraper using Python
A web scraper used to scrap Amazon but could potentially scrap other sites depending on their  robots.txt file. It collects the title of the product you are looking, its price and the date when scraped. It is set to a 24 hour timer while also creating and updating its csv file you can access. Optionally you can send yourself an email when the price reaches below a threshold to buy said product.